find regions only around newly modified cells
changed grid::open to return each cell opened. Added infrastructure to keep track of cells opened in solver (modified cells). find_base_regions only used the newly modified cells to open. Apply open also used info of each cell opened to speed up.

+++ rc_coord.h
@@ -45,9 +45,9 @@ struct rc_coord_hash {
     size_t operator()(rc_coord rc) const {
         std::hash<unsigned> my_hash;
         if(sizeof(unsigned) >= 4)
-            return my_hash((unsigned int) (rc.row << 16) | rc.col);
+            return my_hash((rc.row << 16) | rc.col);
         else
-            return my_hash((unsigned int) (rc.row << 8) | rc.col);
+            return my_hash((rc.row << 8) | rc.col);
     }
 };
